Determine version dynamically from git tags
Removes the requirement to have version defined explicitly in `setup.cfg`/`pyproject.toml`, instead the former
configuration file is removed completely and the version is set dynamically from `pyproject.toml`.

Switches the installation to use the `build` package across all architectures and for the building of a source based
distribution and the version is now determined dynamically on build from the `git tag` (which should be applied when
making a release).